[WIP] Jit64: Add RegCacheAutounlocker for PPC regs
*WIP WARNING*: This is work in progress; interface may change.

This introduces a helper class to be used throghout the JIT to help
with register locking/unlocking.  It maintains a BitSet for all locked
registers, and automatically unlocks registers when going out of scope.

Usage is trivial: substitute calls to RegCache::Lock() to
RegCache::Autolock().  This calls Lock() and returns a new
RegCacheAutounlocker.  The class has a large name on purpose, so that
it becomes idiomatic to write `auto unlocker = gpr.Autounlock(a, b)`.

Should it be necessary to unlock more registers when RCA goes out of
scope, call its AlsoUnlock() method, passing the register index as
usual.  This updates the BitSet.

This currently works only on PPC registers, but will be expanded to
x86 registers in a future commit (possibly fixing this one).

Eventually the new RegCache::Unlock() will be private and
RegCacheAutounlocker and RegCache will be friends. Even better,
RegCache might even use a BitSet to store register locking information.
Who knows?